moved main loop in Window class and renamed Main to Launcher
