+++ app/build.gradle
@@ -32,5 +32,5 @@ java {
 
 application {
     // Define the main class for the application.
-    mainClass = "Main.Main"
+    mainClass = "Launcher.Launcher"
 }
